fix: doc changes and test relocation
